Refactors based on jrddunbr's ideas.
